Готово, фикшу database_api

+++ flask_site_host/data/users.py
@@ -5,13 +5,13 @@ from sqlalchemy import orm
 from .db_session import SqlAlchemyBase
 from flask_login import UserMixin
 from sqlalchemy_serializer import SerializerMixin
-from sqlalchemy.dialects.mysql import INTEGER, MEDIUMTEXT, TINYTEXT
+from sqlalchemy.dialects.mysql import BIGINT, MEDIUMTEXT, TINYTEXT
 
 @dataclass
 class User(SqlAlchemyBase, UserMixin, SerializerMixin):
     __tablename__ = 'users'
 
-    id = sqlalchemy.Column(INTEGER,
+    id = sqlalchemy.Column(BIGINT,
                            primary_key=True, autoincrement=True)
     username = sqlalchemy.Column(TINYTEXT, nullable=True)
     photo_url = sqlalchemy.Column(MEDIUMTEXT, nullable=True)
